feat(TWCounters): require login to submit counters, open to all users
- /TWCounters/forms を requireAuth で保護。ADMIN 限定を廃止しログイン済み全員に開放
- POST /api/counters の ADMIN チェックを削除
- /TWCounters ページをセッション対応: 未ログインはログインボタン、ログイン済みは入力フォームボタンを表示
- /TWCounters/login を /login?callbackUrl=/TWCounters/forms へのリダイレクト route として追加
- requireAuth にcallbackUrl を渡しログイン後に元ページへ戻れるよう修正
