+++ src/app/TWCounters/login/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function LegacyTWCounterLoginPage() {
+  redirect("/login?callbackUrl=/TWCounters/forms");
+}
