+++ src/app/api/counters/route.ts
@@ -1,9 +1,23 @@
 import prisma from "@/lib/prisma/prismaClient";
+import { getCurrentUser } from "@/lib/auth/guards";
 import { NextResponse } from "next/server";
 
 export async function POST(request: Request) {
   try {
-    const { allied_leader, allied_characters, enemy_leader, enemy_characters, description } = await request.json();
+    const user = await getCurrentUser();
+    if (!user) {
+      return NextResponse.json(
+        { success: false, error: "ログインが必要です" },
+        { status: 401 },
+      );
+    }
+    const {
+      allied_leader,
+      allied_characters,
+      enemy_leader,
+      enemy_characters,
+      description,
+    } = await request.json();
 
     const counter = await prisma.counter.create({
       data: {
@@ -15,16 +29,22 @@ export async function POST(request: Request) {
       },
     });
 
-    return NextResponse.json({ 
-      success: true, 
-      data: counter 
-    }, { status: 201 });
+    return NextResponse.json(
+      {
+        success: true,
+        data: counter,
+      },
+      { status: 201 },
+    );
   } catch (error) {
     console.error("Error creating counter:", error);
-    return NextResponse.json({ 
-      success: false, 
-      error: "カウンターの作成に失敗しました" 
-    }, { status: 500 });
+    return NextResponse.json(
+      {
+        success: false,
+        error: "カウンターの作成に失敗しました",
+      },
+      { status: 500 },
+    );
   }
 }
 
@@ -32,19 +52,22 @@ export async function GET() {
   try {
     const counters = await prisma.counter.findMany({
       orderBy: {
-        createdAt: 'desc'
-      }
+        createdAt: "desc",
+      },
     });
-    
-    return NextResponse.json({ 
-      success: true, 
-      data: counters 
+
+    return NextResponse.json({
+      success: true,
+      data: counters,
     });
   } catch (error) {
     console.error("Error fetching counters:", error);
-    return NextResponse.json({ 
-      success: false, 
-      error: "カウンターの取得に失敗しました" 
-    }, { status: 500 });
+    return NextResponse.json(
+      {
+        success: false,
+        error: "カウンターの取得に失敗しました",
+      },
+      { status: 500 },
+    );
   }
-} 
+}
